feat: Enable point selection on the map
This commit adds the functionality to enable point selection on the map in the HomeFragment. It introduces a new button called selectPointsButton in the fragment_home.xml layout file. When the button is clicked, the user can select two points on the map, which are then sent to the server using appropriate APIs. This feature allows the user to define specific points of interest on the map.